refactor: extract manifest parsing into separate class
