+++ src/main/kotlin/com.bugsnag.android.gradle/AndroidManifestInfo.kt
@@ -1,0 +1,3 @@
+package com.bugsnag.android.gradle
+
+data class AndroidManifestInfo(var apiKey: String, var versionCode: String, var buildUUID: String, var versionName: String)
